requirements.txt
asn1crypto==0.24.0
backcall==0.1.0
beautifulsoup4==4.6.0
bleach==2.1.3
boto==2.49.0
boto3==1.9.24
botocore==1.12.24
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
confusable-homoglyphs==3.0.0
cryptography==2.1.4
cycler==0.10.0
decorator==4.3.0
diff-match-patch==20121119
dj-database-url==0.5.0
Django==2.0.3
django-appconf==1.0.2
django-import-export==1.1.0
django-reflinks==0.5.0
django-registration==2.4.1
django-storages==1.7.1
django-tinymce==2.7.0
djangorestframework==3.7.7
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
ez-setup==0.9
gunicorn==19.7.1
html5lib==1.0.1
idna==2.6
inflection==0.3.1
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
jdcal==1.4
jedi==0.12.0
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
kiwisolver==1.0.1
lxml==4.2.0
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
more-itertools==4.1.0
mysqlclient==1.3.12
nbconvert==5.3.1
nbformat==4.4.0
ndg-httpsclient==0.4.4
notebook==5.5.0
numpy==1.14.2
odfpy==1.3.6
openpyxl==2.5.8
pandas==0.22.0
pandocfilters==1.4.2
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.1.0
pinax-referrals==3.0.3
pkg-resources==0.0.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5.2
pyasn1==0.4.2
pycparser==2.18
Pygments==2.2.0
pyOpenSSL==17.5.0
pyparsing==2.2.0
python-dateutil==2.7.0
python-decouple==3.1
pytz==2018.3
PyYAML==3.13
pyzmq==17.0.0
Quandl==3.3.0
requests==2.18.4
s3transfer==0.1.13
scikit-learn==0.19.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tablib==0.12.1
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
traitlets==4.3.2
unicodecsv==0.14.1
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==3.3.1
xlrd==1.1.0
xlwt==1.3.0
